+++ src/simmate/apps/badelf/core/__init__.py
@@ -2,6 +2,5 @@
 
 from .badelf import BadElfToolkit
 from .electride_finder import ElectrideFinder
-from .grid import Grid
 from .partitioning import PartitioningToolkit
 from .voxel_assignment import VoxelAssignmentToolkit
