+++ src/simmate/apps/badelf/core/electride_finder.py
@@ -12,8 +12,8 @@ from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
 from scipy.ndimage import label
 from tqdm import tqdm
 
-from simmate.apps.badelf.core.grid import Grid
 from simmate.apps.badelf.core.partitioning import PartitioningToolkit
+from simmate.apps.badelf.toolkit import Grid
 
 
 class ElectrideFinder:
