+++ src/simmate/apps/badelf/core/partitioning.py
@@ -15,7 +15,7 @@ from scipy.signal import savgol_filter
 from scipy.spatial import ConvexHull
 from tqdm import tqdm
 
-from simmate.apps.badelf.core.grid import Grid
+from simmate.apps.badelf.toolkit import Grid
 from simmate.toolkit import Structure
 
 try:
