+++ src/simmate/apps/badelf/toolkit/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+from .grid import Grid
